Create ecobee_mgr.pl
ecobee manager script.
Currently controls a Heat Recovery Ventilator when wired as a dehumidifier.

Current criteria for this script to run is:
Thermostat must be connected to ecobee servers
Thermostat must be in Heat mode
HRV must be present
Dehumdifier (same HRV) must be present
'dehumidifyWhenHeating' mode must be on.

The script will also activate the furnace fan when there is 1 degree Celsius difference (or more) between the internal temperature sensor and a remote sensor. It will turn off (set to Auto with minOnTime 0) when temperature difference is less than 1C.
This is a current limitation that this will only function when dehumidifier mode is available. These two functions could be independent as one doesn't imply the other. This would be a future code change.